+++ NEWS.md
@@ -1,5 +1,8 @@
 # haven 0.2.0.9000
 
+* `write_dta()` now verifies that variable names are valid Stata variables
+  (#132).
+
 * `write_dta()` and `write_sav()` supports writing date and date/times 
   (#25, #139, #145). `write_sav()` also supports writes hms.